requirements.txt
attrs==19.1.0
awscli==1.16.192
boto3==1.9.79
botocore==1.12.79
Click==7.0
coverage==4.5.3
grpcio==1.21.1
grpcio-tools==1.21.1
isort==4.3.21
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pycocotools==2.0.0
pyquaternion==0.9.5
streamlit==0.48.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.29.1
